chore: useFormat refactor
