+++ src/utils/LanguageEnum.ts
@@ -1,0 +1,12 @@
+export enum Language {
+  bash = 'bash',
+  typescript = 'typescript',
+  tsx = 'typescript',
+  jsx = 'javascript',
+  javascript = 'javascript',
+  java = 'java',
+  python = 'python',
+  go = 'golang',
+  html = 'html',
+  css = 'css',
+}
